refactor: Rename view to screen for consistency
